requirements.txt
aiogram
aiohttp==3.6.2
aioredis==1.3.1
python-dotenv
sqlalchemy
asyncpg
gino
